Add structure (application & machine manifest)
Updated structures.
