+++ applications/ExecutionManager/include/application_state_client.h
@@ -1,0 +1,18 @@
+#ifndef APPLICATION_STATE_CLIENT_H
+#define APPLICATION_STATE_CLIENT_H
+
+
+namespace ExecutionManager {
+
+/**
+ * @brief Available application states.
+ */
+enum class ApplicationState
+{
+    kInitializing,
+    kRunning,
+    kShuttingdown
+};
+
+}
+#endif // APPLICATION_STATE_CLIENT_H
